+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Updated `TryFrom<RecordRef> for RecordRefEnum` for versioning of `StatMsg`,
   `SystemMsg`, and  `ErrorMsg`
 - Fixed conversion of `BboMsg` in `TryFrom<RecordRef> for RecordRefEnum`
+- Implemented `AsyncDecodeRecordRef` and `AsyncDecodeRecord` for `AsyncDecoder` and
+  `AsyncRecordDecoder`
 
 ## 0.36.0 - 2025-06-10